+++ Heatmap/Primitives/Sample.cs
@@ -5,9 +5,9 @@ namespace Heatmap.Primitives
     public readonly struct Sample
     {
         public Vector2 Position { get; }
-        public float Value { get; }
+        public double Value { get; }
 
-        public Sample(Vector2 position, float value)
+        public Sample(Vector2 position, double value)
         {
             Position = position;
             Value = value;
