+++ Heatmap/Range/ConstantRange.cs
@@ -4,6 +4,6 @@ namespace Heatmap.Range
 {
     public sealed class ConstantRange : Range
     {
-        public ConstantRange(float min, float max) : base(Math.Min(min, max), Math.Max(min, max)) { }
+        public ConstantRange(double min, double max) : base(Math.Min(min, max), Math.Max(min, max)) { }
     }
 }
